Store OscillatableSpectra with L/E on the y-axis.
Old state files should continue to work identically, printing a warning, but any new predictions made will pick up the new axis style. This should have minimal impact on the sensitivities. I checked one example and that seems to be true, but this is definitely worth confirming.

If this sticks, we can remove all the SetL calls throughout the codebase, as in the L/E world they don't do anything.